requirements.txt
kafka-python==2.0.2
streamlit==1.24.0
streamlit-autorefresh==0.1.2
pandas==2.2.3
sqlalchemy==2.0.20
requests==2.31.0
pymysql==1.1.0
